+++ edison-jobs/src/test/java/de/otto/edison/jobs/domain/MetaJobRunnableTest.java
@@ -11,9 +11,7 @@ import java.util.Optional;
 import static de.otto.edison.jobs.definition.DefaultJobDefinition.cronJobDefinition;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
+import static org.mockito.Mockito.*;
 
 public class MetaJobRunnableTest {